Livekit Voice Agent/requirements.txt
livekit-agents[silero,turn-detector]~=1.2
livekit-plugins-noise-cancellation~=0.2
livekit-plugins-deepgram
livekit-plugins-elevenlabs
livekit-plugins-openai
python-dotenv
pydantic
flask
flask-cors

